Revert "Feature/#001 connect GitHub api"
