--- lib/github_response.dart
@@ -1,14 +1,0 @@
-import 'package:github_repo_search/github_repo.dart';
-
-class GithubResponse {
-  final List<GithubRepo> items;
-
-  GithubResponse(this.items);
-
-  GithubResponse.fromJson(Map<String, dynamic> json)
-      : items = json['items']
-            .map<GithubRepo>((i) => GithubRepo.fromJson(i))
-            .toList();
-
-  Map<String, dynamic> toJson() => {'items': items};
-}
