+++ pubspec.yaml
@@ -13,7 +13,6 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  http: 0.13.4
 
 dev_dependencies:
   flutter_test:
